requirements.txt
aiohttp==3.8.1
pymongo[srv]==3.12.3
motor==2.5.1
Pyrogram==2.0.30
TgCrypto==1.2.3
bs4
pyshortner
heroku3
aiofiles
validators
python-dotenv
errorsolvers
